Create separate SchduleMainView
Extract ScheduleMainView into separate view and inject it to ScheduleViewController
